extract function for testability
