+++ app/src/ui/branches/ci-status.tsx
@@ -4,6 +4,7 @@ import { APIRefState } from '../../lib/api'
 import { assertNever } from '../../lib/fatal-error'
 import * as classNames from 'classnames'
 import { PullRequestStatus } from '../../models/pull-request'
+import { getSummary } from './pull-request-status'
 
 interface ICIStatusProps {
   /** The classname for the underlying element. */
@@ -17,7 +18,7 @@ interface ICIStatusProps {
 export class CIStatus extends React.Component<ICIStatusProps, {}> {
   public render() {
     const status = this.props.status
-    const ciTitle = generateStatusHistory(status)
+    const ciTitle = getSummary(status)
     const state = status.state
 
     return (
@@ -31,19 +32,6 @@ export class CIStatus extends React.Component<ICIStatusProps, {}> {
         title={ciTitle}
       />
     )
-  }
-}
-
-function generateStatusHistory(prStatus: PullRequestStatus): string {
-  const statusCount = prStatus.statuses.length || 0
-
-  if (statusCount === 0) {
-    return prStatus.state.toUpperCase()
-  } else {
-    const successCount = prStatus.statuses.filter(x => x.state === 'success')
-      .length
-
-    return `${successCount}/${statusCount} checks OK`
   }
 }
